+++ src/fido/Vector.h
@@ -70,12 +70,12 @@ public:
 
   int size() const
   {
-    return values.size();
+    return static_cast<int>(values.size());
   }
 
   int numberEntries() const
   {
-    return nonzeroIndices.size();
+    return static_cast<int>(nonzeroIndices.size());
   }
  
   friend Vector operator /(const Vector & lhs, const Vector & rhs);
